requirements.txt
selenium-wire
reppy
psycopg2
urlcanon
url
python-dotenv